Make an interface for IWritingSystemDefinition and use where appropriate
Actually I used it in all the methods of IWritingSystemRepository,
and in places where it seemed natural to change it to get everything
to compile again.

The main motivation for this is to make testing easier. It's also more
consistent with WritingSystemRepository, which does have an interface.

I also went ahead and deleted an obsolete method, since I was going to
have to change its signature anyway.