Updating the icu.net.dlls to make the FLEx build work.
These are the DLLs that TeamCity currently uses to build Palaso.
For the Jenkins build, which isn't using dependencies I think,
it seems to be necessary to check in this version.

The issue is that FLEx uses a different icu.net.dll, with more content,
but to make Palaso work with it the icu.net.dll that Palaso is built
against must have the same assembly version (currently 1.1.49.0) as
the one in the FLEx Distfiles.

Change-Id: I94ce86591d9476603d8429c8c893edcce844f876