Set preedit location when user types
Instead of setting the preedit location when the user clicks in a field or
when a field receives focus we just set a flag and set the preedit location
when the user starts typing. This fixes most of FWNX-1330; there is still a
slight problem when the view slightly adjusts the scroll position. Original
fix came from Steve McConnel.