FLEx LT-15081: Writing Systems Not Saved Properly (partial)
Addresses the problem that Valid Characters are wiped when restarting FLEx.
They were not being read, because the preceding <special/> tag in the LDML
contains extra Version information that is not being read.  These changes
ensure that we seek to the end of a specific node, not just any node, so such
extraneous information does not prevent loading subsequent information.