Reverse Linux hanging due to USB drive enumeration
This change created issues with the standard send and
receive button in WeSay.  By closing this early, any attempts
to check the USB drives beyond the initial call resulted
in threads hung waiting for a reply.  This first of all
prevented backup menu from working beyond the first
time and also resulted in WeSay hanging on exit.  The
subsequent calls to NDesk require the original thread
that opened the connection to listen for replies.  This
change caused the thread to terminate.