Fix SuperToolTip on Linux
SuperToolTips on Linux so far had two problems: the fading in kind of
worked but it never showed with full opacity due to a mono bug
(#19856). The other problem was that it showed the wrong window type -
the window had a title bar and close button instead of a borderless
window.

This change fixes the windowtype problem and refactors the fading logic
to save an extra cycle.

Change-Id: I641544215d8c19e7a39ba0dc2c2fb5a20925371a