Temp: Work-around crash in WeSay with GeckoFx 29
During a backup, a LogBox is created and the Text property is
accessed during ctor chain.  It seems like there has always been an
C# Null Pointer Exception here, but it now it causes a SIGSEGV in
xulrunner29.  This change will keep the C# Null Pointer Exception
from happening and allow WeSay to work while we figure out the
problem with xulrunner29.