Replace exiftool with taglib#
Main reason is that the previous fix for non-ascii characters in
filenames, involving pumping the data into standardInput and reading
from standardOutput, was crashing for many (typically larger) images.

The new code is cleaner, shorter, doesn't require starting a new
process, backwards compatible, and more likely to be cross-platform.

As a bonus two previously disabled unit tests now pass, and it seems
to be picking up more pre-existing metadata.

One downside is that current releases of programs built using the old
exiftool version will not see changes made to the copyright owner field
made by the new version. Details are in the Metadata class comment.

Note that we are using (for now) a custom build of taglib#. It does not
depend on a 3rd party decompression library and copes with metadata
stored in the way the ArtOfReading uses (ImageMagick).