Work around bug in Mono 3 greedy pattern matching
With Mono 3, the DirectoryScanner is selecting to many files (or
not excluding enough files) when using "$(RootDir)/**/subdir/**/*"
pattern. This works around this (based on previous work on clean
target work around). It would be nice to fix the real problem,
but there will be someone that uses the system mono and not
our custom mono to build.