WS-130: Fix label Text
When no language is selected, one of the labels says,
"Save a Coy of the {0} LDML File...".  Replace Coy with Copy.

Note to reviewers, if you can suggest a substitution string for
{0} that should be used when no language is selected, then I
will consider adding that substitution to this fix.  It seems
ugly to see "{0}" in actual UI.  However, after a language is
selected, you can't unselect one.  Maybe the fix to this is to
select the first language when the dialog is opened.