Add option to display full version in AboutBox
When using a DVCS it's useful to add the SHA1 as revision to the version
number so that it easier to see the state of the code that was used to
build. This can be done by using the AssemblyInformationalVersion
attribute which gets used by Application.ProductVersion.

This change adds the option to display the full version number as taken
from Application.ProductVersion. The default behavior doesn't change and
will still show the short version number (MajorVersion.MinorVersion.Build).
To display the full version number the value TRUE can be passed to the
SILAboutBox c'tor.