Make two dialogs public again
There shouldn't be any need for a client in accessing the
ProblemNotificationDialog and ExceptionReportingDialog dialogs
directly but at least SayMore and Solid do, so I'm changing this
back.

A better way for the clients would be to call one of the
ErrorReport.NotifyUserOfProblem() or ErrorReport.Report*() methods
instead of instantiating these dialogs.

I'm not sure if any other project uses WinFormsExceptionHandler
directly but I'm tired of this discussion so I'm changing it back
to the bad design as well.

This change basically backs out commit 7e63d7.

Change-Id: I86aedd9d306a45fffffad729260810f8c2585f51