Display 11 major languages in the language instead of English
The root driver for this change is Bloom's
https://jira.sil.org/browse/BL-700. User choose "French", but
then see "French" on the covers of the book, when really they should
see français. What they see comes from the "Desired Name" field which
clients optionally enable (Bloom enables it).

Our database doesn't tell us what these language names are. But we
now specify it for some 11 languages common in language work.
This "LocalName" is then given precedence over the English name.

As an added benefit, the user can now search for the language using
this "LocalName" form (in addition to any other name, including the
English one). They also see it as the "primary" name,
which seems good, rather than seeing the English name (even if in a
localized version of the app). The English name is still the first
name listed in the "Other Names" column.