WS-273 "Help" link not visible in Linux
With Linux, the Help link does not move as the
screen size is adjusted.  If the original location
specified is not visible on the screen, the link
does not appear.  The table layout panel next to
this one containing the sort results adjusts
correctly. This change makes the parameters
associated with the two panels identical.  The
link now adjusts relative to the panel size in
both windows and linux.