Don't wait forever for disk device number [Linux]
On Linux, determining on which device a file or directory exists
requires either a low-level system call that is too hard to access
from C# (try it if you don't believe me) or running an external
program to get the information.  The latter process is frequently
hanging for some reason even after the program has finished giving
the desired information.  See https://jira.sil.org/browse/BL-771
for a bug report on this behavior.