Improve temp files
Previously we could end up with two temp files with the same name
if the app created a temp file, deleted it (because it wanted only
the name) and then created a second temp file. This change uses a
slightly different approach for providing a temp file without
creating the file. This guarantees that two temp files have
different names.

Change-Id: Id4b6d1352614aec8ea8604a817dc36c76bb510dc