Fix #180, Creative Commons Licenses images look terrible in PDF/Print
Previously the images were the size we needed for showing in the dialog box. This change replaces them with the large ones supplied by creativecommons.org, which are 403x141. Then, in order to show them on screen without the ugly default shrinking provided by PictureBox, we use a small subclass of PictureBox with its own OnPaint() to draw a high-quality shrunk version.

I also removed a couple related images that weren't being used at all.