Ignore failing tests on TC
For whatever reason the tests that require a readonly file fail
on TC. It turns out that we set the file to be readonly, but
mono on TC reports the file always as being writable which causes
the tests to fail. This change works around this problem by
setting the SkipOnTeamCity category.

Change-Id: I5fb252862373838874e307d6e695b0801af37e13