Support customization of where Registration settings are stored
Also switches Registration settings to be persisted using our
cross-platform class. The first release using this will not upgrade
old registration settings. The only current client is Bloom, which
is OK with this.
Also adds a better algorithm (modify time of config file) for finding
previous configurations to upgrade from.