Refactor keyboard adaptors
The previous approach with the combined keyboard adaptor that relied
on the IbusKeyboardAdaptor for some parts but not for others turned out
to be very confusing to debug and understand. This change modifies the
logic so that we have at most two keyboard adaptors active. The others
don't get initialized and so don't provide any side-effects.

- Split IKeyboardAdaptor interface into IKeyboardRetrievingAdaptor and
  IKeyboardSwitchingAdaptor
- Add IsApplicable property to IKeyboardRetrievingAdaptor that tells if
  this adaptor can retrieve keyboards on this system
- Modify KeyboardType enum: this is now a flag so that a keyboard
  retriever can have the implementation for both system and OtherIm
  types (as is done by the Cinnamon adaptor)
- Don't crash on Linux Mint when ibus is not installed. When ibus is not
  installed we don't have the schema installed. This caused a silent crash
  trying to create the GSettings object. This change checks if the schema
  is installed before trying to create the GSettings object and thus fixes a
  crash on Linux Mint 17.