Create our own InputContext
Previously we used the current InputContext when using the combined
keyboard adaptor. This no longer works for whatever reason: typing
with an IPA KMFL keyboard doesn't activate the keyboard although the
keyboard indicator shows that the IPA keyboard is active. It still
worked with the PalasoUIWindowsForms.TestApp but it turns out that there
we call CreateInputContext before the
KeyboardController.CombinedKeyboardHandling property was set, so we
didn't actually use the current InputContext. After my refactoring in
the previous commit things no longer worked in the test app because now
the property was set.

This change removes the use of the current InputContext and creates a
new one instead.