Don't set ibus keyboard again when it's already set (LT-16179)
When using the IPA KMFL keyboard we get a commit from IBus after
every character. At least on Wasta 14 this causes a keyboard
activation. The previous code that didn't check if the keyboard
was already set causes the IPA keyboard to lose its context thus
preventing the IPA keyboard from working properly. This change
removes the override of IBusKeyboardAlreadySet() method and uses
the base implementation instead.