Fix backspacing with Korean IBus keyboard
When IBus didn't handle the key we should let the control deal with
the keypress. This fixes the problem when typing with Korean Hangul
keyboard it was not possible to backspace past the beginning of the
pre-edit. This change makes it possible, although there is still a
slight problem in that when backspacing over the first character
in the pre-edit the display doesn't get updated, but when backspacing
again the first character in the pre-edit and the character before
that get deleted (as it should be, but the visual indication is
wrong).