Fix PalasolImageTests which, appallingly, were broken and [ignored]
They used to work when we used exiftool, but were ignored in case you didn't have exiftool at hand. When we switched to TagLib a couple years ago, apparently these tests got no attention. Now, they are un-ignored and some small fixes made to make them pass with the current incarnation of PalasoImage.

Also did a little renaming for clarity.

Also made the disposal check use an exception instead of an assert.debug so that the failure is more test friendly (but still only in #DEBUG).