Add SyslogLogger, exception handler, and tests
SyslogLogger will only run on Linux (and Linux-like systems such as OS X,
as long as they have a libc library). It uses the syslog() API defined in
libc to log messages to the local syslog program.

SyslogExceptionHandler likewise will only run on Linux, since it relies on
SyslogLogger to log exceptions to syslog.

Both classes are contained in the SIL.Linux.Logging DLL, which will only
build on Linux (i.e., in the Mono build configurations).