Add IKeyboardDefinition.UseNfcContext property
This property is needed in FieldWorks where we have to know whether
to pass NFC or NFD text to a keyboard when it asks for the context.
Most keyman keyboards are designed for NFD whereas other keyboards
expect NFC. This becomes a problem when the application normalizes
the input text. This flag allows to distinguish the two types of
keyboards (by making the simplified assumption that all keyman
keyboards require NFD). See LT-16637.