Fix GetAvailableLanguageTags()
commit["commit"]["author"]["date"] returns an object that looks
different depending on the current object. This can cause
DateTime.Parse to fail. The better way is to let Newtonsoft.Json
do the conversion to a DateTime.

This fixes a failing test on my machine and probably also problems
on an end user's machine.