Remove obsolete GetExactLanguageMatch
Recent changes totally messed up this method, which has no
unit tests. It's original purpose is obsolete as the table to
which it provided access is now public. It was made for Bloom
which no longer uses it. Rather than try to add tests to something
probably unused, or risk having it break again, I decided to just
remove it.
A version I _think_ is correct may be found in the code removed
by this change. If it is reinstated, unit tests should be provided.