Allow to set prefix for logging
This change adds a new overload for Init() that takes a prefix for
the log file. This is needed for FieldWorks where a new process gets
started when a new project gets opened. Without this change the log
file of the previously started instance would be overwritten when
opening a new project.

Also a little bit of cleanup (moving a method and static variable
around that accidentally got added in the IDisposable region).