BL-2966 Make image library handle image orientation
Image metadata can contain an Orientation field which
indicates how it needs to be flipped and/or rotated to
look right. The image library was ignoring this, which
caused images to look wrong in the toolbox, and was not
copying the metadata to new image files, which caused
them to look wrong elsewhere.
This patch fixes both these problems, though fixing
the first made it unnecessary to fix the second at least
to solve THIS issue.