Allow to append to log file
This is useful when running tests where we might init and shutdown
the logger multiple times, but don't want to overwrite the log file
each time.

Also fix a bug where passing a different logfilePrefix didn't
change the name of the logfile if the name was set previously.

Fix a bug where the WriteEventCore call in the c'tor failed because
the singleton variable wasn't set yet.

(Duplicates commit e1fb51ed141356828c62262075616917b03aa676)