Improve handling of empty or corrupt user.config file
Previously, if the user.config file was corrupt, it would try to put up a UI. However the client may trigger this before UI is even possible. This was killing Bloom silently. This PR adds some unit tests around these situations, verifying that it can self-heal. It also copies corrupt files to ".bad" for debugging purposes. Finally, it adds an optional method a client can use to safely check and heal the settings before UI is possible (Bloom will use this).