Normalize line endings when writing XML files
The previous code assumed that all line endings are written as \r\n.
However that was not the case when the data that we wrote contained
line endings. This change forces line ending normalization so that
line endings always end up as \r\n even when in the data we write.
That helps with FW and LfMerge S/R where the ParserParameters
contain line endings which will otherwise introduce a difference
between Windows and Linux.

This change also allows to pass `false` when calling
`CreateXmlWriterSettings`. This will keep the previous behavior and
is needed for Lift where one test expects that the line endings stay
as they are.