Warning message ignored for 32-bit build and 64-bit build
The C# error is described here:
https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/compiler-messages/cs1591

The project is set up to generate documentation in XML format
which requires that each type and each member have a comment
for the documentation. Not all types and members have been
given comments. This is how the code was for 32-bit.
The warning was being ignored.

When I created the 64-bit release build, I didn't have this
error being ignored and so the release build (which is what
Team City uses) is failing for lack of complete comments.