[API] Mark XmlUtils.AppendAttribute as obsolete
The preferred way is to use SetAttribute which deals with both
setting an attribute and replacing an existing one.

AppendAttribute always appended the attribute to the end of
the collection (and the collection removed an already existing one
with the same name). SetAttribute will keep an existing attribute
and just change its value, thus keeping the order of attributes.

Since this is an API change we also increment the minor version
number.