[linux] Fix location detection for RAMP (LT-16894)
RAMP installs in /opt/RAMP/ramp. LocateInProgramFiles() previously
didn't check subdirectories when running on Linux and so didn't find
RAMP even when installed. This change optionally checks in subdirectories.

This kind of changes API (or at least behavior) on Linux, but this
probably doesn't have noticeable effects for most projects.