Allow to set path for GlobalWritingSystemRepo in unit tests
Get the default path to the GlobalWritingSystemRepo from a field.
This allows unit tests to set the field through reflection so that
they can have the global writing system repo in a defined state.
This is necessary when testing S/R where changes in the global WS
store might cause additional commits to appear.