Replace non-portable use of CultureAndRegionInfoBuilder
This change replaces the use of CultureAndRegionInfoBuilder, which
doesn't exist in Mono, with a call to LanguageLookup. This should make
the code also a bit more robust in the case of some neutral cultures
that it looks like we didn't properly handle previously, e.g. "en".
This change also adds unit tests for the
`WritingSystemFromWindowsLocaleProvider.GetRegion` method which verify
that the method still works in the same way as before (with the
few cases that behave differently mentioned in the comments in the test).