Merge unsigned assemblies (Fixes #865)
We have some unsigned dependencies that we sign on-the-fly during
the build. However, this makes it difficult for clients to use
the library. This change uses ILRepack to merge these dependent
assemblies.

NOTE: there is still a similar problem with ibusdotnet.dll. We can't
merge that because it appears in the API. The solution here will be
to create a nuget package version with signed ibusdotnet.dll.