Add ability to exclude language tags with scripts
For example, when asked to choose a language, users can be confused by
seeing "en-Brai", "en-Desr", "en-Dupl", and "en-Shaw" in addition to "en"
in the list.  See https://issues.bloomlibrary.org/youtrack/issue/BL-7641.
The dialog has a means to manually add a script to the tag if it is
needed.