Use SLDR template when making new WS (LT-20022)
When we save a writing system, we are supposed to preserve any data we don't care about, typically from the SLDR. If the WS has already been saved in the destination folder, we merge changes into the previously saved version. New writing systems that are based on an SLDR template are supposed to have a Template field set, which is used to copy the SLDR version before we look for an existing Save. But this only works for the first place we save a new writing system, since we clear it after using it. I'm not sure why we clear it, so hesitated to change that. A more robust fix, anyway, seems to be that if we are writing a new WS and don't already know of a template to copy and there is a corresponding template in the SLDR, just use it.