Fix keyboard switching on Ubuntu 18.04 (fixes #887)
Ubuntu 18.04 changed the way how to switch keyboards. The previous
method of setting a dconf value is deprecated and no longer works.
Instead this is done by a DBus method call.

This change also adds a property to detect if the program runs in
a Gnome Shell, and fixes a bug where the wrong keyboard adaptor was
used in Ubuntu 18.04.

(cherry picked from commit 296705f3061f39363e7747722232f4d097bfcacd)