[keyboarding] Fix keyboard switching in Gnome (LT-20428)
This change fixes keyboard detection and switching in Gnome Shell.
Some fixes for other environments modified the base class which
broke keyboarding in Gnome Shell. Additionally some of the assumptions
were wrong - it's surprising that things worked earlier...

This change is basically a backport of the changes that were done
earlier on the feature/nuget branch.