[SIL.Windows.Forms] Make GtkSharp and Mono.Posix build-only deps
We need the GtkSharp-signed and Mono.Posix packages for building on
Windows, but not at runtime. On Linux these assemblies are available
as part of Mono, but depending on the Windows versions will break
things. This change sets the `PrivateAssets` attribute for the two
Windows packages. This will allow to build on Windows, but will prevent
a package dependency on these two packages.