Fix AcknowledgementsProvider
For some reason (probably newer mono version) we get a different
exception message when the AcknowledgmentsProvider can't load a
file. I can't see a reason why we'd have to check the exception
message, so now we always ignore a BadImageException.

This change also fixes the SWF test app so that it properly works
with Gecko on Linux.