Use more generic environment variable for WS root
- `USER_DATA_HOME` is more generic for re-use in other parts of
libpalaso if found to be helpful.
- The name of `USER_DATA_HOME` is intended to mean that it's the
user's `XDG_DATA_HOME`. Normally there would not need to be any
distinction, but when running in flatpak, the XDG_DATA_HOME is
application-specific. In this situation, the USER_DATA_HOME could be
set (by the application using libpalaso) to the location of the user's
XDG_DATA_HOME that is external to the flatpak.