Changed signature to new overload of SetSettingsProtection to limit it to ToolStriptems, since that is really the only supported type of component (other than Controls).
Added type checking to ManageComponent so that caller will be alerted if attempting to manage unsupported components.
Improved test dialog DialogWithLinkToSettings to a) not crash when the settings link is clicked, b) correctly indicate how to get hidden controls to appear, and c) test calling ManageComponent and new version of SetSettingsProtection.