Adding .NET Standard to the frameworks which SIL.Core.Desktop targets.
This currently does not compile. There are a few problems to fix, the first of which is solved here. There are a couple IO methods that need to be updated (DirectoryUtilities and FileLocator) with some equivalent fix for the new API. Also, support for the USB operations is planned to only be supported on Framework (until someone wants it badly enough to uplift it to .NET Standard). NDesk.DBus targets Framework 3.5, which is too old to be compatible.

The problem this commit does fix is the old asynchronous pattern used in BasicCommand. This class now uses the recommended pattern (async/await) supported by the .NET Standard platforms. It uses Task.ConfigureAwait(false) to avoid deadlocks with any synchronously blocking (older style) calling code.