Another attempt to fix l10n.proj
During a regular build (of a .csproj) the build system sets
`TargetFramework` for each framework in `TargetFrameworks`. This causes
the `PkgL10NSharp_ExtractXliff` to get defined (in
`obj\l10n.proj.nuget.g.props`). However, if `TargetFrameworks` is set,
the definition of the property gets an additional condition on
`TargetFramework` so that the property isn't set when `UpdateCrowdin`
runs.

This change sets `TargetFramework` (and sets `TargetFrameworks` to the
empty string) to work around this problem.

See also https://github.com/NuGet/Home/issues/11721 and
https://github.com/NuGet/Home/issues/11725