Replaced usage of Mono RelaxNG validator in SIL.Lift with Tenuto.
This was necessary as we move away from the Mono dependency. The RelaxNG package reference only supports .NET Framework, which prevents applications that consume SIL.Lift from targeting .NET Core, etc. Tenuto, though old enough it doesn't have a Nuget package, was easy to convert to .NET Standard. In the process of refactoring I was able to remove progress reporting, as it was confirmed to be almost certainly wholly unused. Validator.GetAnyValidationErrors() was also confirmed to be unused and is now private. This all allowed the cleanup of a good bit of tech debt (e.g. duplicated progress models) in Validator.cs.