Build SIL.Core for net6.0 in addition to netstandard2.0 and .NET 4.6.1
This allows to debug into SIL.Core code from a depending project that is
compiled with net6.0.

Also modify some methods to work around `Assembly.CodeBase` calls. This
method is deprecated in net6.0. In .NET Framework there was a distinction
between `CodeBase` and `Location` where the former returned the original
location of the file and the latter the possibly shadow-copied location
where the assembly actually got loaded from. Shadow-copying often happened
when running unit tests with NUnit. It seems net6 no longer supports
shadow-copying, so we can use `Location` now, but still need to use
`CodeBase` for net461 where we need to access the original location.

+semver:minor