Fix `GetDirectoryDistributedWithApplication`
When running under net6.0 our previous detection of unit tests no
longer works because `GetEntryAssembly()` now returns the test
runner. This change additionally searches the directory where
SIL.Core.dll is located.

This change also moves the list of subdirectories where we look
for files or directories into a private field. When looking for
directories we now also look directly in the application directory,
which is a breaking change from previous behavior where we only
looked in `DistFiles`, `common`, and `src` subdirectories. This
matches the behavior of `GetFileDistributedWithApplication`.

+semver:major