keyboarding: latin layouts and ibus ver via flatpak-spawn
This information should be collected from the host, not from a flatpak
container running the application. This change only uses `gsettings`
when in flatpak and retains the GIO library usage when running not in
a flatpak environment. How to specify the behaviour, given the two
situations, could use some design.

+semver:fix