Checking if Sldr is initialized before each test in SldrTests
This may possibly fix some test flakiness with Sldr. The initialize is also present in the TestEnvironment, but this isn't constructed in every test. By initializing in the SetUp, we guarantee it is. Certainly tests outside this class have been flaky, but I haven't seen any failures in 4-5 runs with this fix.