.NET8 for SIL.Lift.Tests
This included replacing NMock2 with Moq and rewriting almost the entire ParserTests class. The only downside of the port was the loss of ordering verification. I think it's possible to use MoqSequence to verify order, but it seems less flexible than what we were doing in NMock2.

I also changed the base classes for the three matcher classes to their model classes under test, given that Matcher was removed. Those classes become less necessary now, but I don't see a pressing need to refactor them away.