Marked StringCatalog as Windows-only
Also marked its test class with the same. Also marked another test class Windows-only, though it doesn't have to be Windows-only (just how it was written). This attribute is only supported with .NET5 and later, so on Framework there will be no marking.

These changes fix a large number of build warnings.