Moved signed Keyman lib's to their own Nuget package
This fixes a bug with the .NET8-Windows version, where the Keyman lib's are included in the .exe working directory, but never included in the deps.json. This prevents any app from loading them and launching.

You could theoretically remove the deps.json to allow the default .NET assembly loading behavior (e.g. look in the working directory), but the icu.net reference requires a deps.json in the consuming application. Also that probably isn't a very good solution anyway.